refactor(tools): extract edit_overwrite and edit_replace into tools/ (#1190)
